Introduce nav-app-bar and nav-paper website builder components
