+++ libs/navigation/website/src/index.ts
@@ -1,5 +1,8 @@
 export * from './lib/navbar/navbar'
 export * from './lib/navbar/navbar-container'
+export * from './lib/navbar/nav-paper'
+export * from './lib/navbar/nav-app-bar'
 
 export * from './lib/footer/footer'
 export * from './lib/footer/footer-container'
+export * from './lib/link-to-url'
