+++ libs/ui/src/index.ts
@@ -8,3 +8,4 @@ export * from './lib/rating/rating'
 export * from './lib/text-field/text-field'
 export * from './lib/theme'
 export * from './lib/typography/typography'
+export * from './lib/hooks'
