+++ libs/website/builder/src/index.ts
@@ -11,6 +11,8 @@ export * from './lib/footer.interface'
 export * from './lib/image.interface'
 export * from './lib/membership.interface'
 export * from './lib/navbar.interface'
+export * from './lib/nav-app-bar.interface'
+export * from './lib/nav-paper.interface'
 export * from './lib/page.interface'
 export * from './lib/peer.interface'
 export * from './lib/richText.interface'
