+++ libs/website/src/lib/website.provider.tsx
@@ -73,7 +73,7 @@ import {
   SubscriptionList,
   SubscriptionListItem
 } from '@wepublish/membership/website'
-import {Footer, Navbar} from '@wepublish/navigation/website'
+import {Footer, NavAppBar, Navbar, NavPaper} from '@wepublish/navigation/website'
 import {Page, PageSEO} from '@wepublish/page/website'
 import {PeerInformation} from '@wepublish/peering/website'
 import {RenderElement, RenderLeaf} from '@wepublish/richtext/website'
@@ -161,6 +161,8 @@ export const WebsiteProvider = memo<WebsiteProps>(({children}) => (
           Banner={PageBanner}
           PeerInformation={PeerInformation}
           Navbar={Navbar}
+          NavAppBar={NavAppBar}
+          NavPaper={NavPaper}
           Footer={Footer}
           Event={Event}
           EventSEO={EventSEO}
